@@ -6,5 +6,14 @@
 //
 
 extension Never: AppScene {
-    public var scene: Never { fatalError() }
+    public var body: Never { fatalError("Never has no body") }
+
+    @MainActor @preconcurrency
+    public static func _makeView(
+        _ scene: _AppSceneNode<Never>,
+        inputs: _SceneInputs
+    ) -> _SceneOutputs {
+        // For Never, just forward inputs as outputs as per default behavior.
+        _SceneOutputs(appWorlds: inputs.appWorlds)
+    }
 }
